the beginnings of decoding
we successfully decode the serverhello message.
now we need to decode the certificate message.

+++ src/tls.h
@@ -21,7 +21,7 @@ struct ClientHello {
     /* The ID of a session the client wishes to use for this connection.
      * This field is empty if no session_id is available, or if the
      * client wishes to generate new security parameters. */
-    uint16_t session_id_length = 0;
+    uint16_t session_id_length = 0; // should be only 8 bit and cypher_len should be 16 instead??!
     uint8_t session_id[0]; // <0..32> meaning 0 to 32 bytes.
     uint8_t cipher_suites_length = 34;
     uint8_t cipher_suites [34] = { 
@@ -52,7 +52,7 @@ struct ClientHello {
 struct ServerHello {
     uint8_t server_version[2];
     Random random;
-    uint16_t session_id_length;
+    uint8_t session_id_length;
     uint8_t session_id[32]; // FIXME the length of this should be dependant to session_id_length
     uint8_t cipher_suite[2]; // negotiated cypher to use
     uint8_t compression_method;
